BEGINRELEASENOTES
- Added loading pythonizations from podio

ENDRELEASENOTES

Load podio pythonizations (AIDASoft/podio#570) on `import edm4hep`