@tmadlener 2024-09-06
There seems to be some interference with the python (test) modules from EDM4hep? 
 https://github.com/key4hep/EDM4hep/actions/runs/10737553638/job/29779263141?pr=290#step:3:831