@jmcarcell 2024-09-06
But that is the nightlies without https://github.com/AIDASoft/podio/pull/663 included