@m-fila 2024-09-06
Yeah, I think I messed up something with walking through submodules and the local modules are included :unamused: Let me fix that