@tmadlener 2024-09-09
Looks like AIDASoft/podio#667 fixed one issue in the pythonizations and we can properly import now, but we now have a different issue:

https://github.com/key4hep/EDM4hep/blob/b99ba9f9359f9194ec15ab92f5f9c1c422a6d9b3/scripts/createEDM4hepFile.py#L88-L89

Seems to assume a `const MCParticleCollection`, hence calling `operator[](unsigned) const` which returns a `MCParticle` instead of the `operator[](unsigend)` which would return a `MutableMCParticle`. I am not sure whether any of the pythonizations triggers this, in principle this *should work*.