@tmadlener 2024-09-09
These workflows already pick up the latest version of podio as they build it on the fly. The problem in this case really seems to be that the wrong overload for `operator[](unsigned)` is chosen. Looking at #358 the writing seems to work as expected, at least the information can be read back.