@m-fila 2024-09-09
Sorry for the confusion, you are absolutely right. The culprit seems to be the "collection subscript" pythonization that replaces the `__getitem__` (by default bound by cppyy to `operator[]`) with call to `at`. For some reason cppyy chooses to use non-cost overload of `operator[]` and const overload of `at`