@tmadlener 2024-09-10
@m-fila, I will make changes in #361, such that this should start working after it has been merged.