@tmadlener 2024-09-10
At least for our script the changes are not too bad, I think. See:

https://github.com/key4hep/EDM4hep/pull/361/commits/bec3a96390b3b4c4dd7c4f1e45d0683d9e275239