@m-fila 2024-09-13
I timed some usage with and without pythonizations. For commands I list the first time it was called during a session as some things are cached. `edm4hep` local, `podio` from nightlies

| Command                          | REPL    | Without pythonizations (s) | With pythonization (s) |
| -------------------------------- | ------- | -------------------------- | ---------------------- |
| `import edm4hep`                 | python  | 0.6                       | 2.7                    |
| `import edm4hep`                 | ipython | 0.5                        | 2.9                    |
| `edm4hep.<TAB><TAB>`             | python  | ~4                         | ~5                      |
| `edm4hep.<TAB><TAB>`             | ipython | ~1                          | ~1                      |
| `edm4hep.MC<TAB><TAB>`           | python  |    ~2                        | ~2                      |
| `edm4hep.MC<TAB><TAB>`           | ipython |    ~1                        | ~1                      |
| `edm4hep.MCParticleCollection()` | python  |       3.2                     | 1.5                    |
| `edm4hep.MCParticleCollection()` | ipython |               3.2             |             1.5           | 
| `createEDM4hepFile.py`           | python  | 11.6                       | 11.6                   |

The increase in import time is due to the `from podio.pythonizations import load_pythonizations` in the `__init___` which takes more than 2 s on its own (same for `import podio`). The decrease in the object creation time is due to this import rather than `load_pytonizations()`

I don't know how to programmaticaly time autocompletion in python so the estimates for these are very rough

Otherwise I don't see big differences with and without the pythonizations. This might change in the future if there will be more of them

For the sake of benchmark changed `createEDM4hepFile.py`: 

```patch
-    for i in range(3):
-        particle = particles.create()
+    mutable_particles = [particles.create() for _ in range(3)]
+    for particle in mutable_particles:
```