@tmadlener 2024-09-16
Thanks for the extensive checking. This looks about as expected, at least for me. Curious that creating a collection becomes quicker with pythonizations. Could that be because we are measuring different things at that point, i.e. is it possible that part of the things that are done at `edm4hep.MCParticleCollection()` without pythonizations is done during `import edm4hep` with pythonizations?