@m-fila 2024-09-16
Yes, no pythonizations just adding `import podio` to __init__.py also changes the first creation of collection to that ~1.5 s