@jmcarcell 2024-09-16
I wonder if at some point is better to implement the python bindings ourselves...